deepen: absorb OpenRouter credit monitor into daily-report handler
lib/ai/openrouter-credit-monitor.ts exported getOpenRouterCreditStatus
and formatCreditStatusForSlack — a pure fetch + Slack-block formatter
pair — with exactly one production importer: the daily-report Slack
handler's generateHourlySummary path. The shape (extracted for
testability with one caller) matches ADR-0002 and every recent
"inline-shallow-parser-into-Summarize" deepening.

Move both functions and the CreditStatus interface into
daily-report-handler.ts as private, keeping the OpenRouter shape and
its Slack rendering at one address. The credit shape only ever leaves
the module inside a SlackWebhookPayload, so no external caller is
affected.

Deletion test: dropping openrouter-credit-monitor.ts concentrates
complexity at the one caller that owned the credit fetch + rendering.
Nothing else references it (no unit tests existed; the only mention
is a doc comment in __tests__/security/budget-manipulation.test.ts,
updated to point to the new address).

- Delete lib/ai/openrouter-credit-monitor.ts (160 LOC)
- Inline private CreditStatus + getOpenRouterCreditStatus +
  formatCreditStatusForSlack into lib/slack/daily-report-handler.ts
- Update reference comment in
  __tests__/security/budget-manipulation.test.ts

Category 4 (True external): OpenRouter is a third-party service, but
the seam does not need adapters — the daily-report handler is the
only consumer and no test mocks the fetch. Kept as a single deep
function that produces a domain result.
