+++ __tests__/security/budget-manipulation.test.ts
@@ -44,7 +44,7 @@ describe('Budget Manipulation Security Tests - DEPRECATED', () => {
       // OpenRouter handles credit limits via:
       // 1. HTTP 402 errors when credits exhausted
       // 2. /api/v1/auth/key endpoint for credit status
-      // See: lib/ai/openrouter-credit-monitor.ts
+      // See: getOpenRouterCreditStatus in lib/slack/daily-report-handler.ts
       expect(true).toBe(true);
     });
   });
